+++ aether-core/src/chat.rs
@@ -2,18 +2,24 @@ use std::time::Instant;
 
 use azalea::app::{App, Plugin};
 use azalea::chat::{ChatReceivedEvent, SendChatEvent};
+use azalea::ecs::prelude::*;
 use azalea::entity::metadata::Player;
 use azalea::entity::LocalEntity;
 use azalea::prelude::*;
-use azalea::ecs::prelude::*;
 use azalea_anti_afk::config::AntiAFKConfig;
 use azalea_anti_afk::AntiAFK;
 use azalea_utility::auto_eat::{StartAutoEat, StopAutoEat};
 use azalea_utility::auto_totem::AutoTotem;
+use bevy_discord::bot::serenity::all::ChannelId;
+use bevy_discord::bot::DiscordBotRes;
+use bevy_discord::runtime::tokio_runtime;
+use serde_json::json;
+use tracing::error;
 
 use crate::command::AetherCommand;
 use crate::commands::pearl::load::LoadPearl;
 use crate::config::{Bot, Config};
+use crate::discord::{DiscordChannelId, DiscordChatRelay};
 use crate::utils::{parse_chat_content, InWorld};
 
 pub struct ChatPlugin;
@@ -25,15 +31,16 @@ impl Plugin for ChatPlugin {
 }
 
 #[allow(clippy::complexity)]
-fn handle_chat(
+pub fn handle_chat(
     mut events: EventReader<ChatReceivedEvent>,
     query: Query<(&Bot, Option<&InWorld>), (With<Player>, With<LocalEntity>)>,
     config: Res<Config>,
+    discord_bot_res: Option<Res<DiscordBotRes>>,
     mut commands: Commands,
     mut send_chat_event: EventWriter<SendChatEvent>,
     mut load_pearl: EventWriter<LoadPearl>,
     mut start_auto_eat: EventWriter<StartAutoEat>,
-    mut stop_auto_eat: EventWriter<StopAutoEat>
+    mut stop_auto_eat: EventWriter<StopAutoEat>,
 ) {
     for ChatReceivedEvent { entity, packet } in events.read() {
         for (state, in_world) in query.iter() {
@@ -87,13 +94,56 @@ fn handle_chat(
                 });
                 commands.entity(*entity).insert(AutoTotem);
                 start_auto_eat.send(StartAutoEat {
-                    use_inventory: true
+                    use_inventory: true,
                 });
-            } else if content == "You have lost connection to the server" {
-                commands.entity(*entity).remove::<InWorld>();
-                commands.entity(*entity).remove::<AntiAFK>();
-                commands.entity(*entity).remove::<AutoTotem>();
-                stop_auto_eat.send(StopAutoEat);
+
+                if let Some(discord_bot) = &discord_bot_res {
+                    if let Some(chat_relay_channel_id) = state.chat_relay_channel_id {
+                        // Add `DiscordChatRelay` Component
+                        commands.entity(*entity).insert(DiscordChatRelay {
+                            channel_id: ChannelId::new(chat_relay_channel_id),
+                        });
+                    }
+
+                    if let Some(channel_id) = state.channel_id
+                        && let Some(http) = discord_bot.get_http()
+                    {
+                        // Add `DiscordChannelId` Component
+                        commands.entity(*entity).insert(DiscordChannelId {
+                            channel_id: ChannelId::new(channel_id),
+                        });
+
+                        // Send a Notification that bot has joined
+                        let channel_id_clone = channel_id.clone();
+                        let state_clone = state.clone();
+                        tokio_runtime().spawn(async move {
+                            if let Err(e) = http
+                                .send_message(
+                                    ChannelId::new(channel_id_clone),
+                                    Vec::new(),
+                                    &json!({
+                                        "embeds": [
+                                            {
+                                                "title": "Bot Connected!",
+                                                "description": format!("Username: {} \n Role: {:?}", &state_clone.username, &state_clone.role)
+                                            }
+                                        ]
+                                    }),
+                                )
+                                .await
+                            {
+                                error!("Unable to send joined notification to discord, {:?}", e);
+                            }
+                        });
+                    }
+                } else if content == "You have lost connection to the server" {
+                    commands.entity(*entity).remove::<InWorld>();
+                    commands.entity(*entity).remove::<AntiAFK>();
+                    commands.entity(*entity).remove::<AutoTotem>();
+                    stop_auto_eat.send(StopAutoEat);
+                    commands.entity(*entity).remove::<DiscordChatRelay>();
+                    commands.entity(*entity).remove::<DiscordChannelId>();
+                }
             }
         }
     }