+++ aether-core/src/main.rs
@@ -1,9 +1,12 @@
+#![feature(let_chains)]
+
 mod chat;
 mod client;
 mod command;
 mod commands;
 mod config;
 mod config_res;
+mod discord;
 mod plugins;
 mod utils;
 
@@ -12,13 +15,13 @@ use std::time::Duration;
 
 use azalea::{prelude::*, swarm::prelude::*};
 use azalea_anti_afk::AntiAFKPlugin;
-// use azalea_discord::chat_bridge::DiscordChatBridgePlugin;
-// use azalea_discord::DiscordPlugin;
 use azalea_task_manager::TaskManagerPlugin;
 use azalea_utility::UtilityPlugin;
+use bevy_discord::bot::serenity::all::{GatewayIntents, OnlineStatus};
+use bevy_discord::bot::serenity::gateway::ActivityData;
+use bevy_discord::bot::{DiscordBotConfig, DiscordBotPlugin};
 use plugins::AetherDefaultPlugins;
 use tracing::info;
-// use azalea_discord::log_bridge::{DiscordLogBridge, Level};
 
 use crate::config::{Bot, Config, Mode};
 use crate::config_res::ConfigResourcePlugin;
@@ -27,11 +30,8 @@ use crate::config_res::ConfigResourcePlugin;
 async fn main() {
     let config: Config = Config::default();
 
-    // if let Some(webhook) = config.log_bridge {
-    //     DiscordLogBridge::init(webhook, Level::INFO);
-    // }
-
     let server_url: String = config.server.clone();
+    let mut discord_bot_activity = String::from("Bots under control: \n");
 
     let mut swarm = SwarmBuilder::new()
         .set_handler(handle)
@@ -41,8 +41,6 @@ async fn main() {
         .add_plugins(TaskManagerPlugin)
         .add_plugins(UtilityPlugin)
         .add_plugins(AetherDefaultPlugins)
-        // .add_plugins(DiscordPlugin)
-        // .add_plugins(DiscordChatBridgePlugin)
         .join_delay(Duration::from_secs(5));
 
     for (bot_username, bot) in config.bots {
@@ -59,7 +57,28 @@ async fn main() {
                 })
         };
 
+        if config.discord_bot_token.is_some() {
+            discord_bot_activity.insert_str(
+                discord_bot_activity.len(),
+                format!("{}\n", &account.username).as_str(),
+            );
+        }
+
         swarm = swarm.add_account_with_state(account, bot);
+    }
+
+    if let Some(token) = config.discord_bot_token {
+        let discord_bot_config = DiscordBotConfig::default()
+            .activity(ActivityData::playing(discord_bot_activity))
+            .gateway_intents(
+                GatewayIntents::GUILD_MESSAGES
+                    | GatewayIntents::MESSAGE_CONTENT
+                    | GatewayIntents::DIRECT_MESSAGES,
+            )
+            .token(token)
+            .status(OnlineStatus::Online);
+
+        swarm = swarm.add_plugins(DiscordBotPlugin::new(discord_bot_config));
     }
 
     swarm