+++ aether-core/src/plugins.rs
@@ -1,6 +1,6 @@
 use azalea::app::{PluginGroup, PluginGroupBuilder};
 
-use crate::{chat::ChatPlugin, commands::AetherCommandsPlugin};
+use crate::{chat::ChatPlugin, commands::AetherCommandsPlugin, discord::AetherDiscordPlugin};
 
 pub struct AetherDefaultPlugins;
 
@@ -9,5 +9,6 @@ impl PluginGroup for AetherDefaultPlugins {
         PluginGroupBuilder::start::<Self>()
             .add(ChatPlugin)
             .add(AetherCommandsPlugin)
+            .add(AetherDiscordPlugin)
     }
 }